Tests

+++ .scripts/stylesheets/handle-scss.mjs
@@ -1,54 +1,47 @@
+import { existsSync } from 'node:fs';
+import { mkdir, writeFile } from 'node:fs/promises';
 import { dirname, sep } from 'node:path';
-import { writeFile } from 'node:fs/promises';
-import Autoprefixer from 'autoprefixer';
-import CssNano from 'cssnano';
-import rtlcss from 'rtlcss';
-import fsExtra from 'fs-extra';
 import Postcss from 'postcss';
 import Sass from 'sass';
+
+import { Autoprefixer, CssNano, rtlcss } from './configs/css.mjs';
 import { logger } from '../utils/logger.mjs';
 
-const { ensureDir } = fsExtra;
-
-async function handleScssFile(file) {
-  const cssFile = file.replace(`${sep}scss${sep}`, `${sep}css${sep}`)
-    .replace(`${sep}media_source${sep}`, `${sep}media${sep}`)
-    .replace('.scss', '.css');
-
-  let compiled;
-  try {
-    compiled = Sass.compile({ file });
-  } catch (error) {
-    logger(error.formatted);
-    process.exit(1);
+/**
+ * @param { string } inputFile
+ * @param { string } outputFile
+ */
+async function handleScssFile(inputFile, outputFile) {
+  if (!existsSync(inputFile)) {
+    throw new Error(`File ${inputFile} doesn't exist`);
   }
 
+  if (!existsSync(dirname(outputFile))) {
+    await mkdir(dirname(outputFile), { recursive: true, mode: 0o755 });
+  }
+
+  const compiled = Sass.compile(inputFile);
   const plugins = [Autoprefixer];
-  if (cssFile.endsWith('-rtl.css')) plugins.push(rtlcss);
+  if (outputFile.endsWith('-rtl.scss')) plugins.push(rtlcss);
 
   // Auto prefixing
-  const cleaner = Postcss(plugins);
-  const res = await cleaner.process(compiled.css.toString(), { from: undefined });
+  const res = await Postcss(plugins).process(compiled.css.toString(), { from: inputFile });
 
-  // Ensure the folder exists or create it
-  await ensureDir(dirname(cssFile), {});
   await writeFile(
-    cssFile,
-    res.css,
-    { encoding: 'utf8', mode: 0o644 },
+    outputFile,
+    res.css.toString(),
+    { encoding: 'utf8', mode: '0644' },
   );
 
-  const cssMin = await Postcss([CssNano]).process(res.css, { from: undefined });
+  const cssMin = await Postcss([CssNano]).process(res.css.toString(), { from: inputFile });
 
-  // Ensure the folder exists or create it
-  await ensureDir(dirname(cssFile.replace('.css', '.min.css')), {});
   await writeFile(
-    cssFile.replace('.css', '.min.css'),
+    outputFile.replace('.css', '.min.css'),
     cssMin.css,
-    { encoding: 'utf8', mode: 0o644 },
+    { encoding: 'utf8', mode: '0644' },
   );
 
-  logger(`✅ SCSS File compiled: ${cssFile}`);
+  logger(`✅ SCSS File compiled: ${outputFile}`);
 };
 
-export {handleScssFile};
+export { handleScssFile };
